artifacts/legal-luminaire/backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
python-multipart==0.0.20
pydantic==2.10.4
pydantic-settings==2.7.0

# LangChain core
langchain==0.3.14
langchain-openai==0.3.1
langchain-community==0.3.14
langchain-core==0.3.29

# CrewAI multi-agent
crewai==0.86.0
crewai-tools==0.17.0

# Vector DB
chromadb==0.6.3

# Document loaders
pypdf==5.1.0
python-docx==1.1.2
unstructured==0.16.12
pillow==11.1.0
pytesseract==0.3.13

# Google Gemini (Omni-Modal extraction)
langchain-google-genai==2.0.8
google-generativeai==0.8.5

# Web search / verification
tavily-python==0.5.0
requests==2.32.3
beautifulsoup4==4.12.3

# Embeddings
openai==1.59.9
tiktoken==0.8.0

# Utilities
python-dotenv==1.0.1
aiofiles==24.1.0
httpx==0.28.1
redis==5.2.1
scikit-learn==1.5.2
numpy==1.26.4
